+++ lib/observations_gallery_page.dart
@@ -4,7 +4,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:intl/intl.dart';
+
+import 'utils.dart';
 
 /// Page to display the observations as a gallery
 class ObservationsGallary extends StatefulWidget {
@@ -30,7 +31,7 @@ class _ObservationsGallaryState extends State<ObservationsGallary> {
 
     return Column(children: [
       Container(
-        padding: EdgeInsets.symmetric(vertical: 10),
+        padding: EdgeInsets.symmetric(vertical: 0),
         child: Row(
           children: [
             Expanded(
@@ -39,17 +40,14 @@ class _ObservationsGallaryState extends State<ObservationsGallary> {
                 keyboardType: TextInputType.numberWithOptions(
                     decimal: false, signed: false),
                 decoration: InputDecoration(
-                  prefixIcon: Icon(Icons.search_rounded, color: Colors.red),
+                  // prefixIcon: Icon(Icons.search_rounded, color: Colors.red),
                   labelText: "Search Messier",
-                  suffix: SizedBox(
-                    height: 15,
-                    width: 15,
-                    child: IconButton(
-                      padding: EdgeInsets.all(0),
-                      icon: Icon(Icons.clear_rounded, size: 15),
-                      onPressed: () =>
-                          setState(() => messierSearchController.clear()),
-                    ),
+                  suffix: IconButton(
+                    padding: EdgeInsets.all(0),
+                    iconSize: 20,
+                    icon: Icon(Icons.clear_rounded),
+                    onPressed: () =>
+                        setState(() => messierSearchController.clear()),
                   ),
                 ),
                 onChanged: (value) => setState(() {}),
@@ -61,17 +59,14 @@ class _ObservationsGallaryState extends State<ObservationsGallary> {
                 keyboardType: TextInputType.numberWithOptions(
                     decimal: false, signed: false),
                 decoration: InputDecoration(
-                  prefixIcon: Icon(Icons.search_rounded, color: Colors.red),
+                  // prefixIcon: Icon(Icons.search_rounded, color: Colors.red),
                   labelText: "Search NGC",
-                  suffix: SizedBox(
-                    height: 15,
-                    width: 15,
-                    child: IconButton(
-                      padding: EdgeInsets.all(0),
-                      icon: Icon(Icons.clear_rounded, size: 15),
-                      onPressed: () =>
-                          setState(() => ngcSearchController.clear()),
-                    ),
+                  suffix: IconButton(
+                    padding: EdgeInsets.all(0),
+                    iconSize: 20,
+                    icon: Icon(Icons.clear_rounded),
+                    onPressed: () =>
+                        setState(() => ngcSearchController.clear()),
                   ),
                 ),
                 onChanged: (value) => setState(() {}),
@@ -83,20 +78,17 @@ class _ObservationsGallaryState extends State<ObservationsGallary> {
                 keyboardType: TextInputType.datetime,
                 readOnly: true,
                 decoration: InputDecoration(
-                  prefixIcon: Icon(Icons.search_rounded, color: Colors.red),
+                  // prefixIcon: Icon(Icons.search_rounded, color: Colors.red),
                   labelText: "Search by Date",
-                  suffix: SizedBox(
-                    height: 15,
-                    width: 15,
-                    child: IconButton(
-                      padding: EdgeInsets.all(0),
-                      icon: Icon(Icons.clear_rounded, size: 15),
-                      onPressed: () => setState(() {
-                        clickOnClear = true;
-                        dateSearchController.clear();
-                        dateSearchRange = null;
-                      }),
-                    ),
+                  suffix: IconButton(
+                    padding: EdgeInsets.all(0),
+                    iconSize: 20,
+                    icon: Icon(Icons.clear_rounded),
+                    onPressed: () => setState(() {
+                      clickOnClear = true;
+                      dateSearchController.clear();
+                      dateSearchRange = null;
+                    }),
                   ),
                 ),
                 onChanged: (value) => setState(() {}),
@@ -115,9 +107,9 @@ class _ObservationsGallaryState extends State<ObservationsGallary> {
                   );
                   if (dateSearchRange != null)
                     setState(() => dateSearchController.text =
-                        DateFormat.yMMMd().format(dateSearchRange.start) +
+                        dateSearchRange.start.yMMMd +
                             " - " +
-                            DateFormat.yMMMd().format(dateSearchRange.end));
+                            dateSearchRange.end.yMMMd);
                   else
                     setState(() => dateSearchController.clear());
                 },