+++ lib/utils.dart
@@ -1,0 +1,42 @@
+import 'package:intl/intl.dart';
+
+/// Convert decimal degrees to string degree-minute-second format
+String decimalDegreesToDMS(num numeric, String latOrLong) {
+  bool isNegative = false;
+  if (numeric < 0) {
+    isNegative = true;
+    numeric = -numeric;
+  }
+  int degree = numeric.toDouble().floor();
+  int minute = ((numeric - degree) * 60).toDouble().floor();
+  double seconds = (((numeric - degree).toDouble() * 60) - minute) * 60;
+
+  return "$degree\xb0 $minute\' ${seconds.toStringAsFixed(1)}\" " +
+      (latOrLong == 'lat'
+          ? (isNegative ? "S" : "N")
+          : (isNegative ? "W" : "E"));
+}
+
+extension CapExtension on String {
+  /// convert first word to uppercacse
+  String get inCaps => '${this[0].toUpperCase()}${this.substring(1)}';
+
+  /// convert first letter of every word to uppercase
+  String get capitalizeFirstofEach =>
+      this.split(" ").map((str) => str.inCaps).join(" ");
+}
+
+extension CheckExtension on DateTime {
+  /// Check if between [startDate] and [endDate]
+  bool isBetween(DateTime startDate, DateTime endDate) =>
+      this.isBefore(endDate) && this.isAfter(startDate);
+
+  /// Extract the date from a [DateTime] instance
+  DateTime get date => DateFormat.yMd().parse(DateFormat.yMd().format(this));
+
+  /// Get [DateFormat.yMMMd] representation of the date
+  String get yMMMd => DateFormat.yMMMd().format(this);
+
+  /// Get [DateFormat.Hm] representation of the time
+  String get hourMinute => DateFormat.Hm().format(this);
+}